Unify matchers types
