+++ packages/react-day-picker/src/types/CustomModifiers.ts
@@ -1,4 +1,4 @@
-import { Matcher } from './Matcher';
+import { Matcher } from './Matchers';
 import { Modifier } from './Modifier';
 
 /** A map of matchers to use as custom modifiers.*/
