--- packages/react-day-picker/src/types/DateAfter.ts
@@ -1,5 +1,0 @@
-/**
- * Represent a matcher to match a day falling after the specified date, with the
- * date not included.
- */
-export type DateAfter = { after: Date };
