--- packages/react-day-picker/src/types/DateBefore.ts
@@ -1,5 +1,0 @@
-/**
- * Represent a matcher to match a day falling before the specified date, with
- * the date not included.
- */
-export type DateBefore = { before: Date };
