--- packages/react-day-picker/src/types/DateInterval.ts
@@ -1,5 +1,0 @@
-/**
- * Represent a matcher to match a day falling before and after two dates, where
- * the dates are not included.
- */
-export type DateInterval = { before: Date; after: Date };
