--- packages/react-day-picker/src/types/DateRange.ts
@@ -1,5 +1,0 @@
-/**
- * Represent a matcher to match a range of dates. The range can be open.
- * Differently from [[DateInterval]], the dates here are included.
- */
-export type DateRange = { from: Date | undefined; to?: Date | undefined };
