--- packages/react-day-picker/src/types/DayOfWeekMatcher.ts
@@ -1,5 +1,0 @@
-/**
- * Represent a matcher to match a date being one of the specified days of the week (`0-7`, where `0`
- * is Sunday).
- */
-export type DayOfWeek = { dayOfWeek: number[] };
