+++ packages/react-day-picker/src/types/index.ts
@@ -2,17 +2,12 @@ export * from './CaptionLayout';
 export * from './ClassNames';
 export * from './Components';
 export * from './CustomModifiers';
-export * from './DateAfter';
-export * from './DateBefore';
 export * from './DateFormatter';
-export * from './DateInterval';
-export * from './DateRange';
 export * from './DayClickEventHandler';
 export * from './DayFocusEventHandler';
 export * from './DayKeyboardEventHandler';
 export * from './DayLabelFormatter';
 export * from './DayMouseEventHandler';
-export * from './DayOfWeekMatcher';
 export * from './DayPickerProps';
 export * from './DayPickerCustomProps';
 export * from './DayPickerMultipleProps';
@@ -23,18 +18,13 @@ export * from './DayTouchEventHandler';
 export * from './InternalModifier';
 export * from './Formatters';
 export * from './isArrayOfDates';
-export * from './isDateAfterType';
-export * from './isDateBeforeType';
-export * from './isDateInterval';
-export * from './isDateRange';
 export * from './isDateType';
-export * from './isDayOfWeekType';
 export * from './isDayPickerCustom';
 export * from './isDayPickerMultiple';
 export * from './isDayPickerRange';
 export * from './isDayPickerSingle';
 export * from './Labels';
-export * from './Matcher';
+export * from './Matchers';
 export * from './Modifier';
 export * from './ModifierClassNames';
 export * from './Modifiers';
