--- packages/react-day-picker/src/types/isDateAfterType.ts
@@ -1,6 +1,0 @@
-import { DateAfter } from './DateAfter';
-
-/** Returns true if `value` is of type [[DateAfter]]. */
-export function isDateAfterType(value: unknown): value is DateAfter {
-  return Boolean(value && typeof value === 'object' && 'after' in value);
-}
